+++ src/graphics.h
@@ -4,10 +4,10 @@
  * Copyright (c) Steven P. Goldsmith. All rights reserved.
  */
 
-extern void (*setPixel)(uchar *, ushort, uchar);
-extern void (*clearPixel)(uchar *, ushort, uchar);
-extern void (*drawLineH)(uchar *, ushort, uchar, ushort, uchar);
-extern void (*drawLineV)(uchar *, ushort, uchar, ushort, uchar);
+extern void (*setPixel)(uchar *, ushort, ushort);
+extern void (*clearPixel)(uchar *, ushort, ushort);
+extern void (*drawLineH)(uchar *, ushort, ushort, ushort, uchar);
+extern void (*drawLineV)(uchar *, ushort, ushort, ushort, uchar);
 extern void drawLine(uchar *bmp, int x0, int y0, int x1, int y1, uchar setPix);
 extern void drawBezier(uchar *bmp, int x0, int y0, int x1, int y1, int x2,
         int y2, uchar setPix);
